Adding header links
This will make it easier to copy a link to a header.

I did also refactor the post-processors so we easily can add more.
