+++ Makefile
@@ -39,7 +39,7 @@ website/template/assets/app.css: website/node_modules website/assets/*
 website/node_modules: website/package.json
 	cd website && npm install
 
-website-highlight: website/node_modules
+website-post-process: website/node_modules
 ifneq (,$(wildcard .couscous/generated/index.html))
-	cd website && npm run highlight ${PWD}/.couscous/generated;
+	cd website && npm run post-process ${PWD}/.couscous/generated;
 endif
