+++ website/assets/style/main.scss
@@ -55,6 +55,13 @@ h4, .h4 {
     font-size: 1.15rem;
 }
 
+.headerlink {
+    font-size: 80%;
+    font-weight: 400;
+    padding-left: 7px;
+    text-decoration: none;
+}
+
 .content {
     width: 100%;
     padding: 0;
