+++ website/post-process/walk.js
@@ -1,0 +1,31 @@
+const fs = require('fs');
+const path = require('path');
+
+/**
+ * Read directory and all its contents
+ * @see https://stackoverflow.com/a/5827895/1526789
+ */
+let walk = function(dir, done) {
+    var results = [];
+    fs.readdir(dir, function(err, list) {
+        if (err) return done(err);
+        var pending = list.length;
+        if (!pending) return done(null, results);
+        list.forEach(function(file) {
+            file = path.resolve(dir, file);
+            fs.stat(file, function(err, stat) {
+                if (stat && stat.isDirectory()) {
+                    walk(file, function(err, res) {
+                        results = results.concat(res);
+                        if (!--pending) done(null, results);
+                    });
+                } else {
+                    results.push(file);
+                    if (!--pending) done(null, results);
+                }
+            });
+        });
+    });
+};
+
+module.exports = walk;
